service/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.5.3
python-dotenv==1.0.0
python-jose[jwt]==3.3.0
httpx==0.25.2
cryptography==41.0.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
email-validator==2.1.0
python-multipart==0.0.6
google-auth==2.25.2
google-auth-oauthlib==1.1.0
redis 